Fix preformatted merge_request howto
The howto should make it easy to merge changes made in another git
repository. For that reason it shows what commands must be called and
includes the names of branches and url of the repositories inside
eRuby tags. In this case -%> was chosen as special railties version.
-%> will trim the next following whitespace including a newline. This
means that the next git command will be shown at the same line as the
last command and make it impossible to use copy and paste for the
whole block when doing the merge.