Let merge requests live their own life in a branch
Requests are now living inside a ref/reviews branch in the tracking repository, and can be updated though normal pushes to the real repository. 

No authorization yet on pushing updates to a review - merge request owners will not yet be able to push to the target repository, but users with commit access will.