Re-enable breadcrumb icon for teams
The idea of breadcrumbs is give the user a quick overview of where
he's at in the site, so the icons should be easily recognizable.

Having custom icons for teams and users in the breadcrumbs negates
this effect, as the user has to parse the image and realize that
it's a custom icon, which then could mean either a user or a team.

For this reason we keep the icons in the breadcumb standardized.