Remove comment-links from repository and commit-log sidebar
These links were refering to comments made to commits. To me it does
not make sense to have a view where we list all comments made on any
random commit in the repository, since those comments are basically
presented without context.

This change only removes the actual links. There's still a route
for repo/comments, but I'm not sure how that relates to the rest
of the comment processing for commits and/or merge-requests.