Add a summary-field to merge-requests
This is ment as a one-line subject/summary of the changes, similar
to the first line in a git commit message, or the subject of an
e-mail (think the subject of a cover-letter for a patch-series).

This leaves the proposal-field for a more complete description
of the changes.

For now i've only renamed the proposal field to "Descripion" in the
locale, but this should perhaps be done all over, in the data model,
etc.