Remove sidebar from merge-requests and move links to the bottom
Since the merge-request page will have diffs in the future we need
to make room for the diff, just like the commit page. The delete
link is now on the bottom with the preview/add comment buttons.

I also added a link to update the merge-request.

Ideally we should rework these two buttons to be part of the rest
of the form for changing status, commenting, etc.