Store affected lines of an inline diff comment correctly
In order to be able to place the comment correctly in the diff we now
store a tuple of [old_linenumber, new_line_number] along with the
number of lines the comment spans. We store both the first and the
last line number tuple so that we can calculate the placement of
things generated outside the DOM tree (such as the InlineTableCallback
diff rendering callback).