xMake the inline diff comment form work in webkit
By not extracting+inserting the #inline_comment_form html into
different places in the DOM, we avoid weird side-effects in webkit
(and get a lot cleaner code too). The diff container (".file-diff")
must now be passed explicitly to the CommentForm.display() method so
that it knows which one to insert the resulting diff with new comment
into.