Change the way reviewers are notified about new merge requests
When a merge request is created, each reviewer has this added as a favorite. In the same process, a FeedItem is created
for the event associated (now created in the model and accessible as #creation_event). Messages are no longer sent to the
reviewers, as this gave a lot of email to the reviewers. Instead, they will now be able to track the creation (and updates) of merge
requests through their newsfeed.

Since notifications are no longer sent to reviewers, the checkbox on Repositories#edit is no longer displayed (it has no effect). It
is still in the database, though.