Correct executability test for the pre-receive hook
The problem was that File.readlink(filename) may return a relative
path, which is then relative to dirname(filename) rather than the
current directory, so the test for executability fails since this
is not taken into account. However, File.executable?() will do the
right thing regarding symbolic links automatically so there is no
need to try to be clever.