Make clone URL show correctly regardless of enabled protocols
This should show the clone URL correctly regardless of whether HTTP
and/or SSH is available. It will also make the input box used to show
the URL as wide as possible. I have tested it with Firefox on Linux,
Firefox on Windows, IE8, and IE8 in compatibility mode (which should
mimic IE7).