Refactored replace_rails_generated_form function and jQuery helper. Renamed to resourceToggler, and made a note in application.js that it is only used for watching/unwatching projects and repositories.
Updated the favourite helper to not generate a bunch of inline JavaScript that is later
removed by JavaScript on the client (i.e. :method => "..."). Instead, the helper only writes
out the data-request-method attribute, and the refactored resourceToggler picks it up and
adds the Ajax toggle functionality.

Added Sinon.JS for stubbing in JavaScript tests.