Fix an issue where the SSH key processor will complain about missing SSHKey objects in the database.
The message sent to the Message queue to have the key added to the key file was sent by a after_create hook in the SSHKey model.
Since this happens inside a SQL transaction, the message will be sent before the transaction is committed to the database. In situations where the
poller has little load, the message could be picked up and processed before the record was available outside the database transaction in the model.
This would cause the processor to complain that the SSHKey with a given ID could not be found.

This commit moves the call to publish the MQ message to the caller, ie. the controller.