Fix keyboard navigation for diff browser
The callback that handles keypresses to implement keyboard navigation
for the diff browser tried to unbind the callbacks whenever an input
field received or lost input. There were two problems with that.

First, the handler for the blur event didn't call the "enable" method,
so the state of the _enable flag wasn't updated and things became
inconsistent.

Second, bind/unbind were called with dynamically generated functions, so
the unbind wouldn't work, because you actually tried to unbind a
different function than the one you originally bound.

Instead of all that bind/unbind magic, the code now binds once, and then
toggles a simple flag to enable/disable that is checked by a guard
clause in the keypress handler.

Signed-off-by: Björn Steinbrink <steinbrink@saltation.de>