Make branch_and_path() search tags before branches
If a tag name begins with a branch name we would previously identify the
branch name and the later part of the tag name would instead be
considered part of the path. E.g., if the branch is named "dev/test" and
the tag is "dev/test/1", then "dev/test/1/src" would previously result
in branch_ref="dev/test" and path="1/src" whereas it will now be
branch_ref="dev/test/1" and path="src".