Added ability to use an LDAP search for user when authenticating using LDAP mechanisms.
Prior to this change, is was only possible to use a simple LDAP
bind to authenticate against an LDAP server.  Additional config
options have been added to allow binding with a known, unprivileged
account, and using this account to search for the DN of a user,
using a nominated attribute.  This is necessary in organisations
that have user accounts that do not conform well to a simple
template, for example spread over multiple functional and/or
regional organisational units.

It should be reasonably straightforward to extend this to use an
anonymous bind for the initial search, should this be required.

In addition, handling of the "encryption" option has changed
based on feedback from Wari Wahab on the mailing list.  If the
encryption option is set to "none", LDAP will default to using
an unencrypted connection.  The usual warnings about auth over
unencrypted connections apply.

In addition, the script/test_ldap_connection has been updated
to provide the actual DN used for final auth binding in the
case of failures, to assist debugging.

Thanks to Craig Andrews for his initial work on LDAP auth.
I pretty much copied the get_dn_of_user method wholesale from
his original work here:
https://gitorious.org/gitorious/mainline/merge_requests/138

Updated unit tests to include testing of search auth.
Remove some trailing whitespace.