Fixed bug #52, now autocompleting multiple commasep. message receivers.
The problem is in (our old version of) the jquery autocomplete plugin -
it has no concept of multiple slugs in autocompleted input fields.

Solved for now by hacking support for multiple commaseparated fields
into autocomplete.js. In the longterm we should probably upgrade to a
newer version of the autocomplete plugin, though.