wrap authenticate() parameters in a credential object
The authentication plugin API previously took a simple username and
password string pair. In preparation for alternative authentication
mechanisms, wrap the username and password inside a generic "credential"
object and pass that to authenticate(). This
Gitorious::Authentication::Credential object can contain a username and
password, but it can also be extended to support other kinds of
authentication.