Support authorization for push with LDAP backend
When Gitorious is configured to use LDAP for authorization,
authorization for push needs to be done differently from
otherwise.

Rather than collecting all users who have push access and seeing if a
given user is included, we start off with the groups a user is member
of (from LDAP) and see if any of these have push access to the
repository. This is (mainly) because the list of users who are member
of a group lives in LDAP, not our database.