Introducing support for Nginx X-Accel-Redirect configuration
Set frontend_server: nginx to instruct Gitorious to use these headers
instead of X-Sendfile headers. Gitorious will use synthetic URIs for
this, which should be configured in Nginx; read the
doc/templates/nginx.sample.yml for details.

Furthermore, please note that git-http cloning is not yet supported
with this setup.