Call setuid/setgid in script/git-proxy
In able to use the git proxy from Upstart (in Upstart versions which
don't support setgid/setuid themselves) without needing an extra
shell (su -c or equivalent) we might as well do this ourselves.

If all these conditions are met:
- RAILS_ENV=production
- We're running as root
- gitorious.yml specifies gitorious_user

we will switch the script to the gid and uid of the specified user.