Introducing $GITORIOUS_ROOT/bin for CLI work with Gitorious
In order to make it easier to run Gitorious CLI tools directly,
without dealing with switching users, running `bundle exec` and so,
we're now shipping some preconfigured scripts inside bin/.

These scripts will load a functional Gitorious environment and (if
running in production mode) require that either the user configured in
gitorious.yml is used or switch to that user before running the
scripts. This means these scripts can be run either as the configured
user (usually "git") or by root (in which case the scripts will switch
to that user.

The first two CLI tools included are rake (which lets you invoke Rake)
and unicorn (which lets you run Unicorn).