Fix the issue with duplicate merge request versions
When creating a new merge request, a separate processor will ensure
the tracking repository exists and create a new version for the merge
request. This causes another push to occur (since doing git push will
trigger those), and the push processor would intercept this push and
create a version.

With this change, the push processor will only process merge request
updates, not deletions (it didn't before this change) and creations.

Bugs like these are hard to find...

Conflicts:
	test/unit/processors/push_processor_test.rb