Remove SshKey#to_key{file_format}
The keyfile format is a property of the SshKeyFile, not the SshKey
itself. Implementation was moved to the correct class, and is now used
like:

    file = SshKeyFile.new("~/.ssh/authorized_keys")
    file.add(SshKeyFile.format(SshKey.first))

The format is kept external to SshKeyFile#add so as to not needlessly
couple SshKeyFile with SshKey - you can still add keys from any
origin (because #add still takes a string).