Add a caching middleware for Dolt actions
This is enabled when either private mode or private repositories is
enabled. For any action rendered by Dolt, it will alter the
Cache-Control header to private as Dolt itself enforces a fairly
aggressive caching scheme.

To benefit from this, simply enable caching in nginx or set up Varnish
in front of Gitorious.