tests: align "private" repo parameter with controller expectations
Technically the "private" parameter is not a part of the "project"
parameter set. In the "new" view for a project, "private" is a separate
parameter that stands on its own.

Ideally, during a new project creation, the controller is supposed to
merge() "private" into a larger hash along with "project". That works in
Ruby 2.0.0, but it does not work properly in Ruby 1.8.7. Since the test
did not properly reflect what the view currently sends the controller,
just fix the test.