Best guess for Repo browser URLs without commas
When the repository browser is hit with a comma-less URL, we're missing
either the path or the ref/oid. This patch makes the browser smarter,
and is able to recognize a missing path, and will guess at a missing
ref.

Email clients that strip off trailing commas from tree URLs that point
to the root path will now get the expected behavior.