tests: use symbols in params
A couple areas in the test suite used strings instead of symbols in the
params() helpers. Be consistent and use symbols everywhere.

This fixes two intermittent test failures in
CreateProjectRepositoryTest. We were inadvertently passing a params hash
with the duplicate keys "name" and :name. The test suite would fail or
pass depending on how Ruby chose to order the "name" and :name keys for
this hash. Using symbols everywhere eliminates the test failures.