Fix edit user resubmit form
By using UseCase to edit users, Rails fails to properly determine the
HTTP verb to use. When the use case fails, the failure block is called
with an instance of UserValidator. When creating the form for this
object, Rails uses POST, when really it should use PUT. This caused a
routing error when resubmitting a failed user edit. Fixed by specifying
:method => :put in the edit form.