Relative repo path for push event without symlinks
When calculating relative paths between symlinked and non-symlinked
paths, a push event would be created with a path that Gitorious could
not resolve to a repository on the receiving end (the push processor).
This caused push events not to be created in environments where e.g. the
repository_base_path was a synlink (as is the case on gitorious.org
currently).