PushProcessor; don't trip on annotated tags
Calculating new commits in a push used to raise exceptions in
repositories with annotated tags, because you can only pass oids to
Rugged::Walker#hide that point to commit objects.

This commit excludes tags from the list of refs to hide from the walker
and adds an annotated tag to the fixture repository.