Enable arbitrary URLs for http(s) cloning
This allows the usage of any URL as the cloning URL used for HTTP(s)
cloning rather than limiting users to git.hostname as the clone URL.
This is useful when wildcard SSL certificates are used and don't support
subdomains.

For example:
    Regular URL: myserver.example.com
    Clone URL: clone-myserver.example.com

Rather than:
    Regular URL: myserver.example.com
    Clone URL git.myserver.example.com

With normal wildcard SSL certificates the latter clone URL would
give errors since it's a subdomain.

To enable this feature see the http_cloning_domain in gitorious.sample.yml