Use newer libdolt
Allows the use of URLs like:
http://localhost/<project>/<repository>/raw/<sha1>:<path>
Where the sha is not a commit sha, but rather a direct blob sha. When
that happens, the path is basically ignored, and the sha is used
directly with rev_parse by libdolt

This was supported by Gitorious 2, but was accidentally removed in
migrating to using Dolt. Missing this feature is causing certain addons,
like Review board, to fail.