Create MR tracking repository in MR processor *synchronously*
Previously the actual repo cloning was scheduled from
MergeRequestProcessor as another background job but processor's code was
assuming that this repository is already there after issuing
CreateTrackingRepositoryCommand. Because there's no simple way to wait
for the completion of other background job from the current job
CreateTrackingRepositoryCommand learned to (optionally) do the cloning
immediately instead of putting it onto the message queue.