Remove diagnostics functionality
Gitorious web app is going to be deployed in Docker containers. That
means it won't have access to the outside of a container and thus won't
be able to reliably probe for status of other processes or other system
information.

Also, it should only act as a web frontend and management solution for
repositories (and related entities). Monitoring is better solved by
existing tools like Nagios, Munin or similar.