Revert "No need for upstream definitions" and "Set DNS resolver for Nginx"
Dynamic resolving in Nginx requires specifying DNS resolver, which in
turn requires a real DNS server on the host (to resolve from /etc/hosts
populated by Docker). This complicates the setup so we revert back to
static (on startup) resolving.

This is problematic for restarting of linked containers where IP of a
dependent container may change.

Next version of Docker (1.4) is supposed to have "fixed network
configuration" ability for containers which will solve the restart
issue.